move subcommands into their own package #2952
